requirements.txt
aiohttp==3.11.18
alembic==1.13.0
aiosqlite==0.21.0
curl_cffi==0.9.0b2
eth_account==0.13.7
eth_typing==5.2.1
fake_useragent==2.2.0
hexbytes==1.3.1
loguru==0.7.3
pyfiglet==1.0.2
python-dotenv==1.1.0
Requests==2.32.3
rich==14.0.0
tenacity==9.1.2
SQLAlchemy==2.0.40
web3==7.12.0
tweepy-self==1.12.0
